+++ src/ai/utils.rs
@@ -1,4 +1,5 @@
-use dialoguer::Select;
+use console::style;
+use dialoguer::{MultiSelect, Select};
 
 use crate::{config::Config, provider::Provider, utils::console::get_select_theme};
 
@@ -9,8 +10,9 @@ pub async fn on_the_fly_change_model(
     cfg: &mut Config,
     active_model_id: Option<String>,
 ) -> Option<Provider> {
-    let models = cfg.get_available_models();
+    let models = cfg.get_available_models().to_owned();
 
+    // Find the index of the active model
     let active_model = if let Some(model) = active_model_id {
         models.iter().position(|(_, m, _)| *m == model).unwrap_or(0)
     } else if let Some(model) = cfg.active_model() {
@@ -22,6 +24,7 @@ pub async fn on_the_fly_change_model(
         0
     };
 
+    // Get the list of models
     let items = models
         .iter()
         .map(|(provider, _, display_name)| {
@@ -36,20 +39,20 @@ pub async fn on_the_fly_change_model(
         return None;
     }
 
-    let Ok(selection) = Select::with_theme(&get_select_theme())
+    // Create a selection dialog
+    let selection = Select::with_theme(&get_select_theme())
         .with_prompt("Select model")
         .items(&items)
         .default(active_model)
         .interact()
-    else {
-        // Exit if the user cancels the selection (e.g., Ctrl+C)
-        std::process::exit(0);
-    };
+        .unwrap_or_else(|_| std::process::exit(0));
 
     let Some((provider_name, model_id, _)) = models.get(selection) else {
         println!("Invalid selection");
         return None;
     };
+
+    cfg.set_model(provider_name.to_owned(), model_id.to_owned());
 
     let provider = cfg.find_provider(provider_name);
     match provider {
@@ -60,4 +63,36 @@ pub async fn on_the_fly_change_model(
         }
         None => None,
     }
+}
+
+pub fn on_the_fly_select_mcp_client(cfg: &mut Config) {
+    let clients = cfg.mcp_clients_mut();
+
+    let items = clients
+        .iter()
+        .map(|client| format!("{} ({})", client.name(), client.version()))
+        .collect::<Vec<String>>();
+
+    if items.is_empty() {
+        println!("{} No MCP servers configured", style("✗").red());
+        return;
+    }
+
+    let defaults = clients
+        .iter()
+        .map(|client| client.is_enabled())
+        .collect::<Vec<bool>>();
+
+    let selections = MultiSelect::with_theme(&get_select_theme())
+        .with_prompt("Enable / disable MCP servers")
+        .items(&items[..])
+        .defaults(&defaults[..])
+        .interact()
+        .unwrap_or_else(|_| std::process::exit(0));
+
+    for (i, client) in clients.iter_mut().enumerate() {
+        client.set_enabled(selections.contains(&i));
+    }
+
+    cfg.save();
 }